absurd
Struct Arena

pub struct Arena<const N: usize = 1024, const C: usize = 30> { /* private fields */ }

N is the initial chunk size (in bytes) C is the maximum number of chunks chunks are allocated with exponential size, starting at N and doubles. the total size is N * (2^C - 1)

Implementations

impl<const N: usize, const C: usize> Arena<N, C>

pub fn new() -> Self

pub fn alloc<T>(&self, value: T) -> &mut T

Allocates a value in the arena and returns a mutable reference to it. Note that Drop glue is not run on values allocated in the arena.

pub fn calloc<T>(&self, value: Vec<T>) -> &mut [T]

pub fn collect<T>(&self, iter: impl ExactSizeIterator<Item = T>) -> &mut [T]

pub fn alloc_uninitialized<T>(&self) -> &mut MaybeUninit<T>

pub fn calloc_uninitialized<T>(&self, count: usize) -> &mut [MaybeUninit<T>]