absurd
Struct AtomicPtrAlike

pub struct AtomicPtrAlike<T, P: PtrAlike<T>> { /* private fields */ }

Implementations

impl<T, P: PtrAlike<T>> AtomicPtrAlike<T, P>

pub fn new(val: P) -> Self

pub unsafe fn swap_with_order(&self, val: P, order: Ordering) -> P

Safety: relaxed ordering may cause invalid pointers

pub fn swap(&self, val: P) -> P

pub fn swap_seqcst(&self, val: P) -> P

pub unsafe fn store_with_order(&self, val: P, order: Ordering)

Safety: relaxed ordering may cause invalid pointers

pub fn store(&self, val: P)

pub fn store_seqcst(&self, val: P)

pub fn into_inner(self) -> P

impl<T, P: PtrAlike<T> + Copy> AtomicPtrAlike<T, P>

pub unsafe fn compare_exchange_with_order( &self, current: P, new: P, success: Ordering, failure: Ordering, ) -> Result<P, P>

Safety: relaxed ordering may cause invalid pointers

pub fn compare_exchange(&self, current: P, new: P) -> Result<P, P>

Compare the value and swap if it is equal to the current value

pub fn compare_exchange_seqcst(&self, current: P, new: P) -> Result<P, P>

pub unsafe fn compare_exchange_weak_with_order( &self, current: P, new: P, success: Ordering, failure: Ordering, ) -> Result<P, P>

Safety: relaxed ordering may cause invalid pointers

pub fn compare_exchange_weak(&self, current: P, new: P) -> Result<P, P>

Compare the value and swap if it is equal to the current value, but may fail spuriously

pub fn compare_exchange_weak_seqcst(&self, current: P, new: P) -> Result<P, P>

pub unsafe fn load_with_order(&self, order: Ordering) -> P

Safety: relaxed ordering may cause invalid pointers

pub fn load(&self) -> P

pub fn load_seqcst(&self) -> P

impl<T, S> AtomicPtrAlike<T, Option<S>>
where Option<S>: PtrAlike<T>,

pub fn none() -> Self

pub unsafe fn take_with_order(&self, order: Ordering) -> Option<S>

Safety: relaxed ordering may cause invalid pointers

pub fn take(&self) -> Option<S>

pub fn take_seqcst(&self) -> Option<S>

pub unsafe fn try_insert_with_order( &self, val: S, success: Ordering, failure: Ordering, ) -> Result<(), S>

Safety: relaxed ordering may cause invalid pointers

pub fn try_insert(&self, val: S) -> Result<(), S>

pub fn try_insert_seqcst(&self, val: S) -> Result<(), S>