absurd
Struct BitSet

pub struct BitSet<T: Into<usize> + Copy, S = Vec<u32>>(/* private fields */);

BitSet is a set of usize values. E: Integer + BitOps is the element type of the storage T: Into + Copy is the element type of the set S is the storage type, can either be Vec<E> or [E; N]

Implementations

impl<T: Into<usize> + Copy, S> BitSet<T, S>

pub fn with_storage(storage: S) -> Self

impl<E: Integer + BitOps, T: Into<usize> + Copy> BitSet<T, Vec<E>>

pub fn new() -> Self

pub fn with_capacity(capacity: usize) -> Self

Create a new BitSet with at least the capacity (in bits).

impl<E: Integer + BitOps, T: Into<usize> + Copy> BitSet<T, Vec<E>>

pub fn is_subset_of(&self, other: &Self) -> bool

pub fn is_empty(&self) -> bool

impl<E: Integer + BitOps, T: Into<usize> + Copy, const N: usize> BitSet<T, [E; N]>

pub fn new() -> Self

impl<E: Integer + BitOps, T: Into<usize> + Copy, const N: usize> BitSet<T, [E; N]>

pub fn is_subset_of(&self, other: &Self) -> bool

pub fn is_empty(&self) -> bool