Crate absurd

Macros

align_of

cfor

debug

info

new_usize_type

size_of

task

warn

Structs

Arena

N is the initial chunk size (in bytes) C is the maximum number of chunks chunks are allocated with exponential size, starting at N and doubles. the total size is N * (2^C - 1)

ArrayMap

A map based on fixed-length array O(1) insert, O(n) lookup

ArrayMapConstructor

ArraySet

ArraySetConstructor

ArrayVec

ArrayVecIntoIter

AtomicPtrAlike

BTreeMapConstructor

BTreeSetConstructor

BitSet

BitSet is a set of usize values. E: FullBitPrimitive + Integer + BitOps is the element type of the storage T: Into + Copy is the element type of the set S is the storage type, can either be Vec<E> or [E; N]

BitSetConstructor

HashMapConstructor

HashSetConstructor

ManyTimesCell

MinHeap

RAII

a simple smart cell that have a custom destructure callback attached. T can be () for just making the callback.

ShortestPath

Ticker

a helper struct to print iteration count and the elapsed time

Xorshift32

Traits

BitOps

BitSetView

ExtForIgnore

FullBitPrimitive

Integer

IsTrue

Map

MapConstructor

NewUninitPrimitive

One

PtrAlike

PtrExtForReclaimBox

ReadExt

Real

Set

SetConstructor

Signed

SizedExtForApply

SizedExtForLeak

VecExtForPrimitives

Zero

Functions

binary_search

returns a tightened range (l, r) such that f(l) == false && f(r) == true && r - l <= target_range example: binary_search((0.0, 100.0), 1e-6, |x| x * x * x + x > 5.0) returns (1.51598, 1.51599)

new_uninit_primitive

one

zero

Type Aliases

AtomicBox

AtomicMutRef

AtomicOptionBox

AtomicOptionMutRef

AtomicOptionRef

AtomicRef