Crate absurd

Macros

align_of

cfor

debug

info

new_usize_type

size_of

task

warn

Structs

Arena

N is the initial chunk size (in bytes) C is the maximum number of chunks chunks are allocated with exponential size, starting at N and doubles. the total size is N * (2^C - 1)

ArrayMap

A map based on fixed-length array O(1) insert, O(n) lookup

ArrayMapConstructor

ArraySet

ArraySetConstructor

ArrayVec

ArrayVecIntoIter

AtomicPtrAlike

BTreeMapConstructor

BTreeSetConstructor

BitSet

BitSet is a set of usize values. E: FullBitPrimitive + Integer + BitOps is the element type of the storage T: Into + Copy is the element type of the set S is the storage type, can either be Vec<E> or [E; N]

BitSetConstructor

HashMapConstructor

HashSetConstructor

ManyTimesCell

MinHeap

RAII

a simple smart cell that have a custom destructure callback attached. T can be () for just making the callback.

ShortestPath

Ticker

a helper struct to print iteration count and the elapsed time

Xorshift32

Traits

AbelianGroup

An AbelianGroup is a Group where the binary operation is commutative

AbelianGroupInplace

BitOps

BitSetView

CommutativeRing

A CommutativeRing is a Ring where the second binary operation is also commutative

CommutativeRingInplace

ExtForIgnore

Field

A Field is a CommutativeRing where the second binary operation has an inverse element

FieldInplace

FullBitPrimitive

Group

A Group is a Monoid with an inverse element, such that a * a.inverse() == identity()

GroupInplace

Integer

IsTrue

Magma

A Magma is a set with a closed binary operation

MagmaInplace

Map

MapConstructor

Module

An R-Module is an Abelian group that can be scaled by elements of a ring R

ModuleInplace

Monoid

A Monoid is a SemiGroup with an identity element, such that a * identity() == a

MonoidInplace

NewUninitPrimitive

One

PtrAlike

PtrExtForReclaimBox

ReadExt

Real

Ring

A Ring is an AbelianGroup with an additional binary operation that is associative, distributive over the first operation, and has an identity element

RingInplace

SemiGroup

A SemiGroup is a Magma where the binary operation is associative

SemiGroupInplace

Set

SetConstructor

Signed

SizedExtForApply

SizedExtForLeak

VecExtForPrimitives

VectorSpace

A VectorSpace is a Module where the scaling ring is a Field

VectorSpaceInplace

Zero

Functions

binary_search

returns a tightened range (l, r) such that f(l) == false && f(r) == true && r - l <= target_range example: binary_search((0.0, 100.0), 1e-6, |x| x * x * x + x > 5.0) returns (1.51598, 1.51599)

new_uninit_primitive

one

zero

Type Aliases

AtomicBox

AtomicMutRef

AtomicOptionBox

AtomicOptionMutRef

AtomicOptionRef

AtomicRef